Rename UnreadIndicatorsTest to FullAppTests
